Move builder package to api
