+++ src/main/java/me/zeroeightsix/fiber/api/annotation/AnnotatedSettings.java
@@ -4,10 +4,10 @@ import me.zeroeightsix.fiber.NodeOperations;
 import me.zeroeightsix.fiber.api.annotation.convention.SettingNamingConvention;
 import me.zeroeightsix.fiber.api.annotation.exception.MalformedFieldException;
 import me.zeroeightsix.fiber.impl.annotation.convention.NoNamingConvention;
-import me.zeroeightsix.fiber.builder.ConfigAggregateBuilder;
-import me.zeroeightsix.fiber.builder.ConfigTreeBuilder;
-import me.zeroeightsix.fiber.builder.ConfigLeafBuilder;
-import me.zeroeightsix.fiber.builder.constraint.AbstractConstraintsBuilder;
+import me.zeroeightsix.fiber.api.builder.ConfigAggregateBuilder;
+import me.zeroeightsix.fiber.api.builder.ConfigTreeBuilder;
+import me.zeroeightsix.fiber.api.builder.ConfigLeafBuilder;
+import me.zeroeightsix.fiber.api.builder.constraint.AbstractConstraintsBuilder;
 import me.zeroeightsix.fiber.exception.FiberException;
 import me.zeroeightsix.fiber.impl.annotation.magic.TypeMagic;
 import me.zeroeightsix.fiber.tree.ConfigGroupImpl;
