+++ src/main/java/me/zeroeightsix/fiber/api/builder/constraint/AbstractConstraintsBuilder.java
@@ -1,4 +1,4 @@
-package me.zeroeightsix.fiber.builder.constraint;
+package me.zeroeightsix.fiber.api.builder.constraint;
 
 import me.zeroeightsix.fiber.constraint.*;
 import me.zeroeightsix.fiber.exception.RuntimeFiberException;
