+++ src/main/java/me/zeroeightsix/fiber/api/builder/constraint/ConstraintsBuilder.java
@@ -1,6 +1,6 @@
-package me.zeroeightsix.fiber.builder.constraint;
+package me.zeroeightsix.fiber.api.builder.constraint;
 
-import me.zeroeightsix.fiber.builder.ConfigLeafBuilder;
+import me.zeroeightsix.fiber.api.builder.ConfigLeafBuilder;
 import me.zeroeightsix.fiber.constraint.CompositeType;
 import me.zeroeightsix.fiber.constraint.Constraint;
 import me.zeroeightsix.fiber.exception.RuntimeFiberException;
