+++ src/test/java/me/zeroeightsix/fiber/impl/annotation/convention/NoNamingConventionTest.java
@@ -1,4 +1,4 @@
-package me.zeroeightsix.fiber.annotation.convention;
+package me.zeroeightsix.fiber.impl.annotation.convention;
 
 import me.zeroeightsix.fiber.impl.annotation.convention.NoNamingConvention;
 import org.junit.jupiter.api.Test;
