+++ src/test/java/me/zeroeightsix/fiber/tree/ConfigQueryTest.java
@@ -1,6 +1,6 @@
 package me.zeroeightsix.fiber.tree;
 
-import me.zeroeightsix.fiber.builder.ConfigTreeBuilder;
+import me.zeroeightsix.fiber.api.builder.ConfigTreeBuilder;
 import me.zeroeightsix.fiber.exception.FiberQueryException;
 import org.junit.jupiter.api.Test;
